Create LevelManager.cs
working level manager. quest cat's second dialogue will show up (and others will remain the same) so long as the d2 box for other cats are the same as their original dialogue in the inspector. The "collect 3 yarn" objective shows up after talking to the quest cat for the first time and disappears after the quest is completed. The "collected x/2" updates after a kitten is retrieved.